app/requirements.txt
tk
Pillow
pandas
matplotlib
numpy
adafruit-circuitpython-bno055

